BREAKING CHANGE - edited to pass linter and removed circular dependencies When we move forward to 5.x.x TS is much better at seeing the circular dep. The only reason we ever injected the service into two places was so we could proxy to the dialogContainer removeDialog...  one option (used here) is to pass a simple callback to the dialog (from the DialogHolder `    _component.closerCallback = this.removeDialog.bind(this); `) so that we can have our simple close() function
Similiarly we delegated the event handle attachments to the wrapper from holder, just so it would need to come back to holder (via the service) just to call close... simply keeping the outside click handlers inside holder avoids a round trip + circular reference.

IF we ship this - all dialogComponents you created which injected the dialogService ONLY to pass it to the super(), just so we could call dialogservice.removedialog will no longer need that DI, simplifying it i feel, and removing needless requirements from the implementer to handle our wiring problems.